Move assert.h to public header
