+++ src/ds/champ_map.h
@@ -3,7 +3,7 @@
 #pragma once
 
 #include "ds/buffer.h"
-#include "ds/ccf_assert.h"
+#include "ccf/assert.h"
 #include "ds/hash.h"
 #include "ds/map_serializers.h"
