+++ src/ds/thread_messaging.h
@@ -4,7 +4,7 @@
 
 #include "ccf/ds/logger.h"
 #include "ccf/ds/thread_ids.h"
-#include "ds/ccf_assert.h"
+#include "ccf/assert.h"
 
 #include <atomic>
 #include <chrono>
