+++ src/kv/generic_serialise_wrapper.h
@@ -3,7 +3,7 @@
 #pragma once
 
 #include "ds/buffer.h"
-#include "ds/ccf_assert.h"
+#include "ccf/assert.h"
 #include "kv_types.h"
 #include "node/rpc/claims.h"
 #include "serialised_entry.h"
