+++ src/node/historical_queries.h
@@ -4,7 +4,7 @@
 
 #include "ccf/historical_queries_interface.h"
 #include "consensus/ledger_enclave_types.h"
-#include "ds/ccf_assert.h"
+#include "ccf/assert.h"
 #include "kv/store.h"
 #include "node/encryptor.h"
 #include "node/history.h"
